+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "python_template"
+name = "stats_explorer"
 version = "0.2.0"
 authors = [{ name = "Austin Gregg-Smith", email = "blooop@gmail.com" }]
 description = "A python package template"
@@ -9,8 +9,8 @@ license = "MIT"
 dependencies = []
 
 [project.urls]
-Source = "https://github.com/blooop/python_template"
-Home = "https://github.com/blooop/python_template"
+Source = "https://github.com/blooop/stats_explorer"
+Home = "https://github.com/blooop/stats_explorer"
 
 [tool.pixi.project]
 channels = ["conda-forge"]
@@ -30,7 +30,7 @@ python = "3.13.*"
 
 
 [tool.pixi.pypi-dependencies]
-python_template = { path = ".", editable = true }
+stats_explorer = { path = ".", editable = true }
 
 [project.optional-dependencies]
 test = [
@@ -48,7 +48,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build]
-include = ["python_template"]
+include = ["stats_explorer"]
 
 # Environments
 [tool.pixi.environments]
