+++ stats_explorer/basic_class.py
